requirements.txt
Pillow==5.1.0
PyYAML==3.12
reportlab==3.4.0
